requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
decorator==5.1.1
ffmpeg-python==0.2.0
filelock==3.12.2
frozenlist==1.3.3
fsspec==2023.6.0
future==0.18.3
gitdb==4.0.10
GitPython==3.1.31
huggingface-hub==0.15.1
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
llvmlite==0.40.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
more-itertools==9.1.0
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numba==0.57.1
numpy==1.24.4
openai==0.27.8
openai-whisper==20230314
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
protobuf==4.23.4
pyarrow==12.0.1
PyAudio==0.2.13
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.15.1
Pympler==1.0.1
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
referencing==0.29.1
regex==2023.6.3
requests==2.31.0
rich==13.4.2
rpds-py==0.8.8
safetensors==0.3.1
semantic-version==2.10.0
setuptools-rust==1.6.0
six==1.16.0
smmap==5.0.0
SpeechRecognition==3.10.0
streamlit==1.24.1
sympy==1.12
tenacity==8.2.2
tiktoken==0.3.1
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch @ https://download.pytorch.org/whl/cu116/torch-1.13.1%2Bcu116-cp39-cp39-win_amd64.whl#sha256=80a6b55915ac72c087ab85122289431fde5c5a4c85ca83a38c6d11a7ecbfdb35
torchaudio @ https://download.pytorch.org/whl/cu116/torchaudio-0.13.1%2Bcu116-cp39-cp39-win_amd64.whl#sha256=e734a618f102baa252c4876a436686a8130116b26055586fa53d03632379f091
torchvision @ https://download.pytorch.org/whl/cu116/torchvision-0.14.1%2Bcu116-cp39-cp39-win_amd64.whl#sha256=4b75cfe80d1e778f252fce94a7dd4ea35bc66a10efd53c4c63910ee95425face
tornado==6.3.2
tqdm==4.65.0
transformers==4.30.2
typing_extensions==4.7.0
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.3
validators==0.20.0
watchdog==3.0.0
yarl==1.9.2
zipp==3.15.0
